refactor : Correct user to auth package(#19))
