+++ BE/src/main/java/com/codesquad/issuetracker/auth/dto/AccessTokenRequest.java
@@ -1,4 +1,4 @@
-package com.codesquad.issuetracker.user.dto;
+package com.codesquad.issuetracker.auth.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
